Databricks-app/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
gunicorn==23.0.0
pydantic==2.10.3
pandas
plotly
databricks-sql-connector
databricks-sdk
python-dotenv
dash-ag-grid